Add bc tps cmd for ioctl
