+++ cli/ioctl/cmd/bc/bc.go
@@ -26,6 +26,7 @@ var BCCmd = &cobra.Command{
 func init() {
 	BCCmd.AddCommand(bcBlockCmd)
 	BCCmd.AddCommand(bcHeightCmd)
+	BCCmd.AddCommand(bcTpsCmd)
 }
 
 // GetChainMeta gets block chain metadata
@@ -43,49 +44,4 @@ func GetChainMeta() (*iotextypes.ChainMeta, error) {
 		return nil, err
 	}
 	return response.ChainMeta, err
-}
-
-// GetBlockMetaByHeight gets block metadata by height
-func GetBlockMetaByHeight(height uint64) (*iotextypes.BlockMeta, error) {
-	conn, err := util.ConnectToEndpoint()
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
-	cli := iotexapi.NewAPIServiceClient(conn)
-	request := &iotexapi.GetBlockMetasRequest{
-		Lookup: &iotexapi.GetBlockMetasRequest_ByIndex{
-			ByIndex: &iotexapi.GetBlockMetasByIndexRequest{
-				Start: height - 1,
-				Count: 1,
-			},
-		},
-	}
-	ctx := context.Background()
-	response, err := cli.GetBlockMetas(ctx, request)
-	if err != nil {
-		return nil, err
-	}
-	return response.BlkMetas[0], err
-}
-
-// GetBlockMetaByHash gets block metadata by hash
-func GetBlockMetaByHash(hash string) (*iotextypes.BlockMeta, error) {
-	conn, err := util.ConnectToEndpoint()
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
-	cli := iotexapi.NewAPIServiceClient(conn)
-	request := &iotexapi.GetBlockMetasRequest{
-		Lookup: &iotexapi.GetBlockMetasRequest_ByHash{
-			ByHash: &iotexapi.GetBlockMetaByHashRequest{BlkHash: hash},
-		},
-	}
-	ctx := context.Background()
-	response, err := cli.GetBlockMetas(ctx, request)
-	if err != nil {
-		return nil, err
-	}
-	return response.BlkMetas[0], err
 }
