+++ cli/ioctl/cmd/bc/bcblock.go
@@ -7,12 +7,15 @@
 package bc
 
 import (
+	"context"
 	"fmt"
 	"strconv"
 
 	"github.com/spf13/cobra"
 
+	"github.com/iotexproject/iotex-core/cli/ioctl/util"
 	"github.com/iotexproject/iotex-core/cli/ioctl/validator"
+	"github.com/iotexproject/iotex-core/protogen/iotexapi"
 	"github.com/iotexproject/iotex-core/protogen/iotextypes"
 )
 
@@ -63,4 +66,49 @@ func getBlock(args []string) string {
 		fmt.Sprintf("Receipt Root: %s\n", blockMeta.ReceiptRoot) +
 		fmt.Sprintf("Delta State Digest: %s\n", blockMeta.DeltaStateDigest) +
 		fmt.Sprintf("Hash: %s", blockMeta.Hash)
+}
+
+// GetBlockMetaByHeight gets block metadata by height
+func GetBlockMetaByHeight(height uint64) (*iotextypes.BlockMeta, error) {
+	conn, err := util.ConnectToEndpoint()
+	if err != nil {
+		return nil, err
+	}
+	defer conn.Close()
+	cli := iotexapi.NewAPIServiceClient(conn)
+	request := &iotexapi.GetBlockMetasRequest{
+		Lookup: &iotexapi.GetBlockMetasRequest_ByIndex{
+			ByIndex: &iotexapi.GetBlockMetasByIndexRequest{
+				Start: height - 1,
+				Count: 1,
+			},
+		},
+	}
+	ctx := context.Background()
+	response, err := cli.GetBlockMetas(ctx, request)
+	if err != nil {
+		return nil, err
+	}
+	return response.BlkMetas[0], err
+}
+
+// GetBlockMetaByHash gets block metadata by hash
+func GetBlockMetaByHash(hash string) (*iotextypes.BlockMeta, error) {
+	conn, err := util.ConnectToEndpoint()
+	if err != nil {
+		return nil, err
+	}
+	defer conn.Close()
+	cli := iotexapi.NewAPIServiceClient(conn)
+	request := &iotexapi.GetBlockMetasRequest{
+		Lookup: &iotexapi.GetBlockMetasRequest_ByHash{
+			ByHash: &iotexapi.GetBlockMetaByHashRequest{BlkHash: hash},
+		},
+	}
+	ctx := context.Background()
+	response, err := cli.GetBlockMetas(ctx, request)
+	if err != nil {
+		return nil, err
+	}
+	return response.BlkMetas[0], err
 }
